feat(M3): Node/Python/PHP 三語言實作＋4×4 跨語言矩陣全綠
- Node（@argonguard/passwords）：@node-rs/argon2 引擎、async API、
  isWellFormed surrogate 攔截、215 測試綠、自寫 PHC vs 原生 hash() 交叉比對
- Python（argonguard-passwords）：argon2-cffi 引擎、3.9 相容、224 測試綠、
  py.typed、typing.Protocol legacy 介面
- PHP（argonguard/passwords）：原生 password_hash＋sodium fallback 雙 provider、
  off-by-1024 專測（memlimit=m×1024 對照 engine-units.json 權威值）、218 測試綠、
  #[\SensitiveParameter]、絕不降級 bcrypt
- 四語言 harness 契約 12/12；HARNESS_CMD 改 repo 相對指令
- 守門 3（matrix.yml）：4×4 round-trip 矩陣（16 lang-pair × 凍結密碼集 × 三檔位
  ＋needs-rehash＋frontier/t=1 邊界）本地 ALL GREEN
- fix(dotnet)：version 欄位不做窄化 cast（SPEC S4 禁止靜默 wrap，
  v=2^32+19 → unsupported_version）＋回歸測試——Node agent 交叉發現

+++ dotnet/src/ArgonGuard.Passwords/Phc/PhcHash.cs
@@ -3,7 +3,7 @@ namespace ArgonGuard.Passwords.Phc;
 /// <summary>嚴格解析後的 PHC 欄位。<c>Algorithm</c> 保留原字串供 dispatch 判斷是否 argon2id。</summary>
 internal sealed class PhcHash
 {
-    public PhcHash(string algorithm, int? version, long m, long t, long p,
+    public PhcHash(string algorithm, long? version, long m, long t, long p,
                    byte[] salt, byte[] tag, bool hasKeyid, bool hasData)
     {
         Algorithm = algorithm;
@@ -18,7 +18,7 @@ internal sealed class PhcHash
     }
 
     public string Algorithm { get; }
-    public int? Version { get; }
+    public long? Version { get; }
     public long M { get; }
     public long T { get; }
     public long P { get; }
